+++ TravelDiscovery/Discover/Destination/DestinationHeaderContainer.swift
@@ -1,0 +1,28 @@
+//
+//  DestinationHeaderContainer.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 08/06/23.
+//
+
+import SwiftUI
+
+struct DestinationHeaderContainer: UIViewControllerRepresentable {
+    
+    typealias UIViewControllerType = UIViewController
+    
+    func makeUIViewController(context: Context) -> UIViewController {
+        let viewController = UIViewController()
+        viewController.view.backgroundColor = .red
+        return viewController
+    }
+
+    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
+    }
+}
+
+struct DestinationHeaderContainer_Previews: PreviewProvider {
+    static var previews: some View {
+        DestinationHeaderContainer()
+    }
+}
