+++ TravelDiscovery/Discover/PopularDestinationsView.swift
@@ -68,20 +68,6 @@ struct PopularDestinationTile: View {
     }
 }
 
-struct Container: UIViewControllerRepresentable {
-    
-    typealias UIViewControllerType = UIViewController
-    
-    func makeUIViewController(context: Context) -> UIViewController {
-        let viewController = UIViewController()
-        viewController.view.backgroundColor = .red
-        return viewController
-    }
-
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
-    }
-}
-
 import MapKit
 
 struct PopularDestinationsDetailsView: View {
@@ -96,11 +82,8 @@ struct PopularDestinationsDetailsView: View {
     
     var body: some View {
         ScrollView {
-            Image(destination.imageName)
-                .resizable()
-                .scaledToFill()
-                .frame(height: 150)
-                .clipped()
+            DestinationHeaderContainer()
+                .frame(height: 250)
             VStack(alignment: .leading) {
                 Text(destination.city)
                     .font(.system(size: 18, weight: .bold))
